requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
mysql-connector-python
requests
tomli >= 1.1.0 ; python_version < "3.11"
